Add changed frameworks testing to main pipeline
This commit adds generative testing of only changed frameworks against
mysql57 and mysql80 to the main pipeline. This ensures that if someone
makes a change to one of the frameworks, that it still passes testing
against mysql (ensuring that the tests themselves are still accurate).
Testing against ReadySet is only carried out in a separate nightly
pipeline.