Implementation of a single-threaded reader for the fastly demo.
This commit adds the `Reader`, a program that simulates multiple
read clients issuing reads to the fastly noria deployment. This
is accomplished by generating new parameters to the parameterized
read query `fastly_read_query.sql`, that queries the ReadySet view
`fastly_view.sql`.

`fastly_read_query.sql` and `fastly_view.sql` represents a simple
version of the read query and view, respectively. This will be
expanded as we support the read query/view in it's entirety.