Derive Arbitrary for query_generator::QuerySeed
Using the derive macro for Arbitrary from the test_strategy crate,
derive an instance of the proptest::Arbitrary trait for
query_generator::QuerySeed and all the various things that make it up.
Currently this isn't used for anything, but going forward this will be
used as the source of randomly fuzz-testing queries.