Lint and validate packer config when it's changed
Previously, the steps to lint and validate the packer config files would
only run when the build packer images step was unblocked - this meant
that we accidentally and inadvertently committed changes to these files
that failed both of these steps. Now, the step to check formatting of
the packer images runs always whenever these files are changed, and the
build_packer_pipeline.sh script takes an ONLY_VALIDATE env var as
configuration that is set by a new step that also runs if the packer
config files are changed.