Save the ID of the last successful build to S3
I misunderstood the Buildkite documentation around build artifacts
originally - it turns out that downloading build artifacts *is* in fact
per build ID, and you have to pass a build to filter by when
downloading. We'd still like the desired behavior of getting the last
successful build of the release binaries, though, so what this does is
add an extra step after the release binaries are successfully built to
upload a single file to a (configurable) S3 bucket containing the ID of
the last successful buildkite build, which is then used as the build ID
when downloading the release binaries to build packer images from.