Add the columns used for pagination to the query graph
Currently ORDER BY can only use columns that are already projected,
columns and expressions that aren't projected can't be used. This change
makes it so that those columns are added to the query graph, and then
can be used for post processing in the `do_read` function.

However we can't simply add columns to the graph, as they will then be
added to the schema and impropely returned by reads. Therefore another
postprocessing step takes place that removes those special columns from
the schema when results are returned.

Fixes #[ENG-248]
Fixes #[ENG-337]