Teach Extremum how to recover from a deleted extremum value
If you deleted the records corresponding to the extremum value reported
by the extremum grouped operator, it would have no idea how to continue
(because it has no way of knowing what the next highest value is without
holding on to all records it's ever received).

This fixes that omission by introducing a new error type
(`GroupedStateLost`) and replaying all records through a grouped
operator when its `apply()` method reports that error type. A logictest
is added to validate this behaviour.

(fixes ENG-341)