Implement DB fallback for msql-srv queries
Customer database fallback is implemented to provide a path for
queries to be sent directy to the customer's mysql server when noria
cannot handle the query. Fallback is implemented at the noria client
to recover from as many readyset errors as possible. The backend
struct reroutes queries that would (or did) go to the noria_connector
to the mysql_connector. Redirection is done when parsing the query
fails or when the noria_connector returns an error when running a query.

This commit breaks conversion to f32 from DOUBLE columns. This was
implemented in mysql_common v0.21.0 after Value::Float(f64) was split
into Value::Float(f32) and Value::Double(f64).
(https://github.com/blackbeam/rust_mysql_common/pull/17#issuecomment-603751038).

Known Issues/ Future Work:
- The backend struct does not implement a postgres connector, so fallback
will only work when a mysql server is present.
- Fallback is only implemented for mysql queries. Parameterized select
  statements, prepares and executes are not implemented.
- The noria client will respond Query Ok instead of Empty Set when a
  fallback query returns no results. This can be avoided by providing a
dummy column to msql-srv, but this workaround is not included.
- Metrics for fallback queries are not implemented.
- MySqlSelect results are not parsed by the js-client
- Fallback testing needs to be implemented