Limit the depth of Arbitrary QuerySeeds
Relying exclusively on the generic Arbitrary implementation for
QuerySeeds was pretty quickly causing test case generation to blow up
with a stack overflow error, due to the self-recursive nesting of
QuerySeed -> Vec<Subquery> -> QuerySeed. This rewrites that to be
an *explicit* Arbitrary implementation using the `prop_recursive`
combinator from proptest, which limits the amount of nesting that
happens to something a lot more reasonable.