Add a proptest-based fuzzer to noria-logictest
Add a new subcommand, `fuzz`, to noria-logictest, which uses a proptest
TestRunner along with the Arbitrary impl for QuerySeed to generate
random vectors of query seeds and, using logic factored out of the
`generate` subcommand, uses those query seeds to generate a TestScript
based on results from MySQL which it then runs against noria.

As much as possible here has been shared between generate and fuzz -
specifically, most of the actual logic for running queries against mysql
and saving the results has been moved to a `run` method on
Seed (returning a TestScript), which both generate and fuzz now use.

Finally, since proptest doesn't have the ability to run tests in an
async context directly, we have moved the initialization of the Tokio
runtime out of the top-level `main` function and into the commands
themselves - in this case, each test we run generates its own runtime.