Adds `UniqueRepeated` and `UniformWithoutReplacement` column specs.
This change adds two new `ColumnGenerationSpec` variants aimed at
generating rows in batches rather than in isolation.

The new  variant `UniqueRepeated` generates the same value for a set
batch of rows before generating a new unique value.
`UniformWithoutReplacement` pulls from a uniform distribution without
replacement, and allows resetting the distribution after a predetermined
batch of rows has been generated. These variants allow more precise
generation of rows to match demo specs as shown by their usage in
the Fastly demo.

The current implementation of UniformWithoutReplacement repeatedly
attempts to generate a random value that we have not seen yet in the
current batch. This can be impossible or take a prohibitive amount
of time if the batch size is larger or near that of the number of
elements in the distribution. However, for our current use cases,
the fastly demo, the number of elements in the distribution is much
larger than the batch size.