Allow randomly generating limits for topk
There are a number of parameters used when generating queries from
various query operations that have much too large of a state space to be
enumerated exhaustively - in these cases, since we're testing
generatively now, we can just add fields to the query operations but
hardcode them when parsing from command line parameters, leaving them to
be generated randomly by proptest via the derived Arbitrary impl for the
type.

This is the first and probably simplest of such cases - `limit` for
topk.