Allow generating random types for aggregates
Similar to what we did for the limit field on topk, add a `column_type`
field to every AggregateType (except for GroupConcat, for which only the
Text column type is valid), hardcoding it in the case of user-provided
operations but generating it using a proptest strategy when randomly
generating test cases. In the case of Sum and Avg, this is limited to
only numeric types, using a new proptest strategy for SqlType.