Allow generating random subquery join operators
Similarly to how we did for aggregate column types and topk limits,
parametrize the subquery position enum variants on a join operator,
hardcode it when parsing from command-line parameters, and allow it to
be generated when generating Arbitrary query operations.

I played around a little bit with putting this on the Subquery
QueryOperation instead of SubqueryPosition itself, but settled on this
way of doing things to fully pave the way for WHERE IN with subqueries,
which don't have a relevant JoinOperator.