noria-logictest: output copyable command listing for newly-passing tests
If a large number of tests are now passing, renaming them all is a pain
(if you aren't familiar with Vim/Emacs keyboard macros that is :p).
Output an easily copy-and-pastable set of commands to rename all the
newly-passing tests, to make generating a commit that marks them all as
passing easier.