Allow generating random values+types for filters
Allow generating random column types and literal values to filter on
when generating the Filter query operation.

This is slightly more complicated because we need to make sure the type
of the column and the type of the value we're filtering on match up - to
accomplish this, we have a new associated function on Literal that
constructs a proptest Strategy to generate an arbitrary literal value
with the given SQLType.