Exit with a non-zero exit status if we find bugs
It's nice to be able to see via the status of the pipeline if the fuzz
testing has found bugs - to achieve that, we need to actually fail the
execution of the binary. We achieve that by returning an anyhow error
via `bail!` rather than just printing that we've found bugs.