Update top-level docstring for QueryGenerator
The scope of the query generator has grown quite a bit since it was
first created to run benchmarks - we now support random in addition to
deterministic data, we have a bunch of new fields on query operations,
we support subqueries, etc. This brings the module-level docstring
up-to-date with all of those changes, giving a more thorough explanation
of what the query generator is being used for *today*, and expanding a
little bit on how it works under the hood.