When collapsing WHERE IN with aggregate query, don't parametrize
Parametrizing WHERE IN queries with aggregates results in a multi row
response, where only one row is expected. For some aggregate we lack the
required context to aggregate the many rows into a single response, so
instead of parametrizing, we simply desugar them to a series of OR or AND.

This desugaring was already supposed to take place in the MIR lowering code,
but the way it works now the code is never reached as the collapsing was
taking place in the rewrite stage anyhow.