Upload proptest regressions as build artifacts
If a test run fails in CI, it *could* be because one of our proptests
found a novel test failure - in this case, it's useful to be able to
look at the seeds for the regressions so we can locally reproduce the
failure. Buildkite exposes all artifacts as downloadable files to users,
so we can achieve this by uploading the regressions files as artifacts
if the tests fail.