Count(*) ignore NULLs
Count(*) needs to simply count rows and ignore nulls. This behavior can
fit into our current method of picking a random over column, as long as
there's an option to count_nulls.

This commit adds that option, pipes it through, and deterministically
avoids nulls depending on whether this condition is set. This feature
gets flipped in the re-write from CountStar -> Count, thereby counting
nulls and returning the correct result for count(*).