Remove Literal::UnsignedInteger
There's no syntactic difference between an unsigned integer and a
positive signed integer, so having a different constructor for a signed
vs an unsigned integer is a little silly - we never even parsed it! This
removes it, which has the extra advantage of getting rid of one way to
trigger the `unimplemented!()` in the From<Literal> impl for DataType.