Make Literal->DataType conversion fallible
Convert the impl for From<Literal> for DataType to TryFrom, since
literal conversion to DataType can currently fail - notably, this
happens if we have a non-utf8 binary blob or a placeholder literal. Both
of those should be fixed - we want to support binary blobs as data
types, and Placeholder should not be a Literal - but in the meantime
it's best to avoid the panics and return actual errors instead.

This also *implements* conversion from Literal::CurrentDate and
Literal::Blob, since neither was previously supported.