Include both alias and origin column name in a view schema
Previously if a column was given an alias in a select statement, then
the view schema would only be able to address the column under that alias
whereas for MySQL it is still possible to use both the alias and the name
of the column.

In this change a more generic ViewSchema struct is introduced, that is more
extensible than the previously used vector, and also more idiomatic.

The ViewSchema includes the same schema as before, but also adds the base
table information for columns for which it is relevant.

Fixes #[ENG-336]