Implement emit empty values for no-group queries
For queries involving no group by, if the result of a COUNT or
GROUP_CONCAT is the empty value for either, that value should be emitted
in the case there is NO group by, which is MySQL behavior.

This adds checks internally to see if we should emit the default empty
value, by checking for the aggregation type, and whether we currently
have anything in the group by clause. This needs to be implemented as
part of the GroupedOperation trait, because at the point of receiving
the Aggregation type, it has lost it's type as being any T that
implements GroupedOperation. That means that we need a method on
GroupedOperation that can type check for us.

This functionally changes how we respond to COUNT and GROUP_CONCAT
queries with no group by. In those cases, we now correctly return the
default zero value for each rather than simply returning empty sets.