Add TODO for empty table case
In the case that we have an empty table that never gets populated with
data, we bail early having never received any deltas. This produces an
empty set for all queries until deltas do arrive.

This TODO in part documents the problem, which I plan to fix in a
different PR.