Logictest: ensure num results returned matches specified
This change checks the number of expected results against the number of
returned results in a logictest and returns a failure if they differ.
This check is already conducted for logictests that return hashed
results.

Previously, if a logictest returned M results but expected N results,
pass/failure would be determined by a comparison of the first min(M,N)
results.

This change causes logictests/param,joins.test to fail.