Reimplement GROUP_CONCAT with slightly cursed hack
Aggregate functions work by deriving the next value from the current
value and a set of diffs. However, in the case of GROUP_CONCAT, this is
imperfect; the "current value" is a string representation of a bunch of
DataTypes, which the old GroupConcat implementation would attempt to do
terrible things with.

Instead of using that pile of hacks, we use a different pile of hacks in
a new, alternative implementation (named GroupConkitten). This
implementation stores the original DataTypes used to generate the output
value in the GroupConkitten struct (which is kinda unkosher!) and
compares the current value to the string representation of the stored
value when an update happens; if they do not match, the GroupedStateLost
mechanism introduced earlier forces the state to be regenerated. We can
then work with the actual data with no loss of fidelity, which seems to
have fixed quite a few problems that plagued the old implementation.