Fix nom-sql index slicing behavior to be safe from panics.
In multiple points we use index slicing that can panic if the
index does not exist. Fixed to instead use safe operations as
most of these cases did not require index slicing.

We previously would not convert from digit to u8 for precision
values and instead converted just the first u8 value in a &str.
This is replaced with actually converting from a &str to the
u8 precision. For instance "255" would not be converted to u8: 255,
it would take the byte value for "2" and use that as the u8 precision.