Create makefile for builds and test generation
This commit creates a basic makefile for building and pushing docker
images. The makefile also includes a command any engineer can use to run
the generated logic tests locally, which should be the same behavior as
the current nightly pipeline. It also includes a command to run the main
test suite using docker-compose, which allows engineers to run the
binlog testing locally.