Add retry on TCP errors logic to logictests
At least on macOS, when running a lot of generated logictests in parallel
the OS runs out of available ports, because TCP sockets linger for sometime
after being closed and ports can't be reused right away. This change adds
a timeout with retry to TCP errors to allow lingering sockets to fully close.