Make Real internal representation just an f64
The mantissa, exponent, sign representation we had switched to is
literally just the bit representation x86 uses internally for floating
points (albeit with the sign at the end, instead of the beginning),
except implemented in userspace without being able to take advantage of
the processor's FPU (meaning unless LLVM is doing some really impressive
magic, it's likely much slower than native float operations). Not to
mention, our epsilon-based Eq implementation didn't match up to the Hash
impl which hashes the raw bits of the floats, opening us up to *all*
sorts of soundness bugs when we use a Real DataType as the key of a
HashMap (something we famously do... *very* frequently).

This refactors the internal representation to just be an f64, removing a
lot of the code that existed to support the manual IEEE 754
re-implementation, and makes the Hash and Eq impls match up by both
being based on the *bits* of the floating point internally.