Optimize reader to support higher query throughput.
The current reader assumes that each thread will return a value
per query interval, but as that interval gets smaller and throughput
gets higher, the ordering queries are processed at ReadySet may
not guarantee that each thread will have a query execute in a given
interval.

Support not specifying a target_qps to issue queries as quickly as
possible.