Make launchpad::Indices return an error
Make the indices and cloned_indices methods in launchpad::Indices return
an error rather than using panicking index operations. Since rust lacks
a trait for a fallible `get` operation, this requires implementing the
trait explicitly on multiple different types rather than just making it
an extension trait off of `std::ops::Index` - ostensibly, this trait
could be implemented for *all* types that implement Index, but rather
than do all of that work here I've left it to only slices, hashmaps, and
btreemaps for now - in the future we may want to add more.

Lots of stuff had to be updated at call sites in noria/dataflow to make
this work - a lot of stuff there is now using `unwrap_or_default` under
the assumption that the records we're receiving either are the correct
length or have already been checked to be the correct length - for any
that aren't (already deeply incorrect behavior) I'm not entirely sure
what'd happen, but it'll definitely be wrong.