Error handling for noria/noria/src/data.rs.
Fixed all the clippy warnings concerning this file.
In particular, this required implementing `TryFrom<String>` for `DataType`,
which caused a cascade of changes.
We are now using `try_from` for the conversion of `String` to `DataType`,
returning a `ReadySetResult` everywhere except in two cases:
1. When we are using `try_from` from a literal (such as
`DataType::try_from("global")`` or `DataType::try_from("")`) where we know the
transformation won't fail and therefore we can use `.unwrap()` (with a `#[allow
(clippy::unwrapped_used)]` and a comment explaining that).
2. In all the files under `noria/applications`, since they are not part of the
product, but rather part of the testing/benchmarking frameworks, where
modifying the methods to return a `ReadySetResult` seemed out-of-scope for the
purposes of the current error-handling efforts.