Implement all of get_parameter_columns_recurse
Implement (as a first take) all of the remaining pattern matches in
noria_client::utils::get_parameter_columns_recurse, with extra care
taken to correctly support BETWEEN (since it is in fact supported as a
top-level condition with parameters). In all other cases, we bottom out
at returning an empty vector of parameter columns if we don't find a
condition that we know how to handle right now (specifically ANDed
together comparisons on columns) - the idea being that later, when we
hit the literals when walking the WHERE expression otherwise we'll
return an unsupported error.