Refactor joins to use a "Side" enum
Rather than passing around booleans which cryptically translate to
true=left, false=right when specifying which parent of a join we're
talking about in various circumstances, refactor joins to use a new
"Side" enum, with variants for Left and Right. This should end up
translating to effectively the same machine code, while keeping things a
lot more immediately readable and self-documenting.

This is a refactor I've been wanting to do for a while now, but I'm
doing it now because I'm about to do something in unions that is going
to use this enum as well, so I figured I'd clean up as I went.