Exclude duplicate rows for ORed together filters
For the sake of enabling future reuse, when converting OR in filters to
dataflow we make each side of the OR its own filter node, and then
combine together those nodes with a union node. This is fine if the
filters are completely disjoint - but our union nodes aren't actually
union, they're closer to "concat", where the same row appearing in both
sides of the union is emittted twice - this means that in the case of
filters which overlap for any reason we'd previously emit the same row
multiple times where it passed both filters on either side of the OR.

This fixes that by adding a new "DuplicateMode" parameter to Union,
which if set to the new "BagUnion" variant will implement the behavior
we're looking loking for for filters, where the *maximum* number of
duplicates of each row is eventually emitted by the union. Internally,
this is implemented with a new "BagUnionState" data structure which for
each unique row received by the union tracks the number of duplicates
emitted for that row, and the parent of the union that contains the
maximum number of duplicates - this algorithm is checked for correctness
using a property test.

One (internal to dataflow) limitation here is that making the
BagUnionState algorithm as efficient as possible required limiting the
number of possible parents for a BagUnion operator to 2 - this is fine,
since for OR-ing together filters we only ever make union nodes with
exactly 2 parents (I'm pretty sure >2 parent union nodes are only ever
currently used as shard mergers).

There's a new hand-written logictest that's been added covering this
specific behavior, plus 88 previously-failing logictests have been
renamed to passing.