Still print errors for expected-failing logictests
If a logictest that is expected to fail fails when running, it's still
nice to see *why* it fails - especially if you're planning on fixing
the bug that's causing it to fail! This commit changes the logictest
runner to still print the error that caused an expected-failing
logictest to fail when we run it and it fails as expected.