TopK column order needs to be the same as parent
Similar to filters, the topk operator is not capable of reordering
columns, so when we add new columns we need to make sure those columns
are in the same position as they would be when added to the parent node.
This was causing queries with both an aggregate and a topk to have
columns out-of-order and hence return the wrong results. As a result, 4
logictests have been moved from expected-failure to passing.