Refactor most checked_sub to saturating_sub
This commit refactors most uses of checked_sub in dataflow to instead
use saturating_sub, where appropriate. This targets places in code where
we unwrap, making the use of checked_sub more or less useless in
production.

saturating_sub does what we want in all of the places addressed in code.
The only thing that would have caused us to have had None with
checked_sub, would be a bug that causes us to subtract memory that we
hadn't previously added. In that case converting to zero is the correct
behavior.