Add a sleep command to logictest
Add a new `sleep` command to logictest, which allows *explicitly*
sleeping for a given number of milliseconds in logictests. I added this
to help with debugging an issue I suspected was due to write propagation
time (it was!) and will subsequently be *automatically* sleeping after
all insert or delete statements - but I figured since I added this
anyway and it feels generally useful I'd commit it.