lookup refactor -> ReadySetResult
This refactors lookup() to return an iterator over items of
ReadySetResult, and rely on the caller to bubble up any errors it sees.

This is necessary because lookup calls query_through which in turn calls
lookup. They both need the same function signatures, and because
iterators are lazy, we can't bubble up the errors earlier (by wrapping
the entire function return in a ReadySetResult, for instance).

As a result of this commit, we can now correctly handle errors caused by
improper filters that users write.

Co-Authored-By: Griffin Smith <griffin@readyset.io>