Always sleep after statements in logictests
It's not just the binlog replication that needs propagation time after
statements in logictests - noria itself is also eventually consistent,
so if eg we have a DELETE followed by some queries we need to wait for
those negative deltas to propagate forwards. This removes the
"needs_sleep" boolean that we previously only set if we were running in
binlog replication mode, instead just *always* sleeping after a
statement.