Allow aggregating over group by col in extremum
This is essentially the same as 6a2a838ff (Allow group by col being
aggregated over., 2021-06-03), except for the extremum operator in
addition to just count. The same rationale for removing this check there
applies here as well - it's totally valid in SQL to do this, and we
return the correct results if we do.

As a result of this change, 50 previously-failing logictests have now
been moved to passing.