Binlog replication support for propagating RYW timestamps.
Previously, the binlog ignored GTID_EVENTS that preceeded
each transaction. Similar to the debezium connector, we now
record the sequence number associated with a GTID_EVENT and
apply it to the following table actions. This allows us to
support RYW consistency by attaching the sequence number
to writes as they are propagated through the system.