Return None if merging no packets in group_commit
In the original version of this function, we returned None in the case
that we received no packets. That functionality was lost through
iteration, even though the function signature never changed.

It appears that is still the correct behavior, so rather than unwrap and
potentially panic, let's deal with the no packets case correctly.

Co-Authored-By: Griffin Smith <griffin@readyset.io>