Move post-lookup order and limit into readers
Previously, the post-lookup operations we did to (re-)order and limit
the resultset of a lookup were specified by noria-client, as part of the
`ViewQuery` struct specifying the lookup. The client itself was expected
to parse the query, extract the columns to order by (and the limit), and
resolve those column names to column indices by inspecting the schema of
the noria View. This approach has broken down in a number of ways - for
one, the source of truth for which columns (really, expressions that may
or may not be aliased) correspond to which is entirely internal to MIR -
once we get to dataflow, the idea is that we should be discarding any
knowledge about the *names* of the columns as much as possible, and this
breaks that boundary. For another, and more saliently at the moment,
figuring out how to order or limit a query in the client prevents
rewrite passes, such as the one that's about to be added to remove
order+limit from queries that can only possibly return one row, from
functioning properly, since those rewrite passes happen internally to
noria-server and noria-client never knows about them.

This commit moves the order_by and limit fields out of ViewQuery, and
into a new `PostLookup` struct that's on both the reader node in the
graph and on the `SingleReadHandle` that's queried as part of handling
view lookups. This is then used inside the readers instead of ViewQuery
to do the actual ordering and limiting (plus post-lookup filters, which
have to remain on the ViewQuery because they're only filled in at lookup
time). In MIR, we have a name-based as opposed to index-based version of
the order_by (same as we do for the rest of MIR) stored on the
MirNodeInner::Leaf.

While we're at it, this also removes the vestigial `operator` that was
on the reader node, which was left over from an earlier, abandoned
approach to range queries and completely unused.