Remove TopK when we can only return one row
TopK on a non-grouped aggregate query is pointless, since the query can
only ever possibly return one row - furthermore, there was a bug where
due to the topk node being added, as part of projecting through the
order column for the topk, we'd add that column to the group_by of the
aggregate node, causing the aggregate node to return incorrect results.

This adds a rewrite pass to remove the ORDER BY and LIMIT clauses
entirely in the case that the query can only possibly return one row,
and also validate the MySQL `only_full_group_by` rules to make sure that
when we're *not* removing the topk, it's properly ordering by either a
grouped column or the result of the aggregate itself.