error handling: Fix domain/mod.rs
For all known sources of panics¹ in domain/mod.rs, either remove the
panicking (by replacing it with a return of Err, in most cases) or
explicitly allow the relevant clippy lint along with documentation of
the reason the panic can't actually happen.

There's a *lot* involved in this, as since domain is quite
optimized (especially to avoid cloning) there's a lot of non-typechecked
invariants that we have to deal with. The most notable ones - documented
in the relevant fields of the domain - are the ones relating to things
like replay paths referencing other relational data structures in
domains - eg every node in a replay path is guaranteed to reference a
node in `self.nodes` in the domain, etc etc. There's also some more
subtle stuff around when we *know* a particular replay path is
partial (as opposed to full), hence guaranteeing that eg it will
definitely have a non-None replay_key. Finally, there's the standard
borrow-checker-mangling that's been done around passing around Options
that we know will never be None, but that we want to be able to do
things like `take()`ing - that's all commented inline

On the side of things that *aren't* actually invariants, but real errors
that seem (to me) like they could actually happen, I've made sure to
validate all received messages at the edge to make sure they eg refer to
real nodes, or refer to actual reader nodes in the case that a message
is only valid for a reader node, etc. Plus, a *real* panic I've seen
happen in the wild, notably the channel_coordinator not being able to
construct a channel for a domain, has been turned into a real error
message (but is not currently handled - that should *definitely* happen
soon as part of general fault-tolerance work).

¹: This is given by running clippy with `-W clippy::unwrap_used -W
   clippy::panic -W clippy::indexing_slicing`