Bootstrap new noria servers with existing domains.
Currently when a noria server starts up, it registers
with the controller and as recipes change, can receive
updates about new domains that are created, or domains
that are moved. However, the server does not know about
domains that were created before they joined the deployment.
As a result, replays fail as they do not know about a domain
in the replay path.

This change makes `GossipDomainInformation` take a vector of
domain descriptors rather than a single domain descriptor and
is passed to new nodes that register with the controller.