Demo support for measuring write propagation time.
The write propagation time demo client uses a RYW ticket
to block read on it's preceeding write. This works because
(1) a reader will not satisfy a read associated with a ticket
until a timestamp packet with a greater sequence number arrives,
(2) a timestamp packet is propagated after a write associated with
that timestamp. As a result, the time it takes to read a key with
a ticket associated with its write is an upper bound on the
write propagation time.

This allows us to estimate an upper bound on the time it takes for
a write to propagate from the mysql instance to the reader node.