Don't group by every column in group_concat
Actually pass in the set of column indices to the constructor for
group_concat, and use that to determine the set of column indices to
group by rather than grouping by all non-aggregated columns in the
parent node. I'm not sure *why* this was the way it was originally -
there was even a *comment* saying that this was something we wanted to
do for this reason - but this is the way things work in mysql (as they
do for all aggregates!).