Don't set max_level_trace in logictests
If we have this feature flag, trace-level log events will still
get *created*, they'll just be filtered out by the runtime log filter.
This means that the log events we have to, for example, log the entire
query graph at various points in time will have to allocate memory for
the graphviz representation of a query graph with potentially thousands
of nodes, significantly slowing down the execution of the generated
logictest suite and, in the worst case, even causing it to OOM. This
removes that feature flag - if we need it for specific cases we can
re-add it, or even add a feature flag of our own to compile with it
enabled.