Fixes for error handling for noria/noria/src/data.rs
The changes made were:
1. We went back to implementing `From<String> for DataType` instead of using
`TryFrom`, which internally uses `TryFrom<&'a [u8]>` (try from bytes) and then
unwraps. It's safe to unwrap there since the only way for the `TryFrom` from
bytes to fail is if the bytes are not a valid UTF-8 string, but since Rust's
Strings are UTF-8 encoded (thus valid), using `TryFrom` the bytes of a String,
is safe. Same goes for `&str`.
2. We went back to using index slicing since there are invariants that guarantee
 that it's safe to use them.