Remove literal quotes from Display impl for text
We had tried to display `DataType` string types with extra quotes
around them. There was a comment left asking if this is what we really
want to be doing, and I think it's not. It's not how strings are
displayed back to you in MySQL, for one, and two it's just extra noise,
that in this particular case also adds extra memory and needless
computation in calculating string weights.