reuse: IS NULL does not imply IS NOT NULL
Because, I assume, comparing against NULL for every other operator
results in always returning NULL, the predicate implication code had a
check to see if a binary operator had NULL on the right-hand side, and
if so would indiscriminately consider the predicates as implying each
other. This is incorrect in the case of both the IS and IS NOT
operators, which *do* have the ability to distinguish between NULL and
non-NULL values.