Mark tests failing due to ENG-427 as out-of-scope
All tests which involve an aggregate and a subquery will eventually put
an aggregate inside of a subquery, and then fail due to
ENG-427 (aggregates in subqueries should emit a single empty row), which
is currently out-of-scope for our v0.1 milestone. We don't want these
tests clogging up our passing exhaustive-tests metric, so this moves
them out of logictests/generated into a new out-of-scope directory,
under a subdirectory named after ENG-427.

I did this entirely by looking at the *names* of the tests, so there may
be several tests that are *currently* failing due to *other* reasons,
but there are so many of them that explicitly looking at the results
would take prohibitively long to do.