Add a pass to detect self-joins on the same column
Add an AST-level pass to the server which doesn't perform any rewrites,
but checks for the specific problematic self-join case in ENG-411 where
the two sides of a join key are eventually dependent on the same column
in the same table on both sides, and returns an unsupported error if it
finds that case.

This ends up being reasonably complex, as we need to trace through
arbitrarily many subqueries and CTEs to find the column, and there are a
number of scoping rules for referring to tables in queries. A number of
tests cover some simple supported and unsupported cases here, but it's
entirely possible I missed something and this check either rejects valid
queries, or passes through queries that don't hit the problematic case -
regardless, it's a starting point, and it should catch at the very least
99% of the problematic queries and return a helpful error message
instead of just returning the wrong results.