Make ReplayPath::path a Vec1
Replay paths can never be empty - previously, this invariant was tested
manually before construction, then we unwrap()ped when taking nodes out
of the path. If we change the datatype to be a Vec1, we can use the
type-system to prove that we checked it can't be empty, and no longer
have to unwrap() at use-time. Nice!